Push the state up into Game instead of each Square.
Game needs to operate on the whole board, and it's better in React to
then keep the whole state there instead of having Game query each
Square component every time it needs to look at the "whole board".

Therefore we need to change around the onClick events to call the one
in Game to set its state.