TTT: Fooling around with new Modals trying to get animations to work
This is mostly a checkpoint commit. But "The Game" still works.

Trying to get a pop-up style Modal, through a Portal to work with
animations. This first step is to rework the Modal class to be
separate from the Portal stuff and to have it support a property to
show/hide itself.

Side effect of this is that it fixed for not the dangling "node has
been removed React node in the old modal version after the reset game
was clicked.