TTT: Better Portal abstraction - pass it a Component type to render
Instead creating a potential C++ style mess of Component class
heirarchy, just pass Portal the React Component type to render inside
the Portal.

Add a stupid Tooltip Component, which is just another kind of Modal
with the first potentially useful name I thought of to test this
Portal functionality with multiple types.

Fixup the CSS of the two types of "modals" so they don't overlap.